rename types to common
