+++ local-volume/provisioner/pkg/common/common.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package types
+package common
 
 import (
 	"github.com/kubernetes-incubator/external-storage/local-volume/provisioner/pkg/cache"
